+++ examples/10-tagged-union/go-purego/tagged_union/purego_test.go
@@ -87,6 +87,7 @@ func TestPuregoTaggedUnion(t *testing.T) {
 
 	assertSignalSnapshots(t)
 	assertTaggedUnionValueParameters(t)
+	assertTaggedUnionValueReturns(t)
 	assertValueVariants(t)
 	assertConcurrentCallsRacingClose(t)
 	assertConcurrentProjectionsRacingClose(t)